refactor: table 스타일 중복 및 history table 적용
